requirements.txt
Django=1.5.2
django-extjs4=1.2.1
django-tastypie=0.10.0
gunicorn=10.0
hammock=0.2.4
requests=1.2.3
six=1.4.1
wsgiref=0.1.2